refactor: consolidate scan API to /scans resource model
Scan creation previously lived under /predict while retrieval, download, and\nreport routes already used the /scans resource model. That split made the\nAPI inconsistent and left legacy router code in place.\n\nThis change introduces a canonical routers/scans.py implementation for\nscan creation, metadata, downloads, and reports, updates the frontend to\nPOST /scans, removes the live /predict route from the app, and keeps thin\ncompatibility modules for existing imports and tests.
